+++ lib/vector-core/src/lib.rs
@@ -29,6 +29,7 @@ pub mod mapping;
 pub mod metrics;
 #[cfg(test)]
 mod test_util;
+pub mod transforms;
 
 pub use buffers;
