+++ src/transforms/concat.rs
@@ -211,7 +211,7 @@ impl FunctionTransform for Concat {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::event::Event;
+    use crate::{event::Event, transforms::test::transform_one};
 
     #[test]
     fn generate_config() {
@@ -235,8 +235,8 @@ mod tests {
             ],
         );
 
-        let new_event = transform.transform_one(event).unwrap();
-        assert_eq!(new_event, expected);
+        let result = transform_one(&mut transform, event).unwrap();
+        assert_eq!(result, expected);
     }
 
     #[test]
@@ -256,9 +256,10 @@ mod tests {
             ],
         );
 
-        let new_event = transform.transform_one(event).unwrap();
-        assert_eq!(new_event, expected);
+        let result = transform_one(&mut transform, event).unwrap();
+        assert_eq!(result, expected);
     }
+
     #[test]
     fn concat_mixed() {
         let mut event = Event::from("message");
@@ -279,8 +280,8 @@ mod tests {
             ],
         );
 
-        let new_event = transform.transform_one(event).unwrap();
-        assert_eq!(new_event, expected);
+        let result = transform_one(&mut transform, event).unwrap();
+        assert_eq!(result, expected);
     }
 
     #[test]
@@ -294,7 +295,7 @@ mod tests {
             vec![Substring::new("only[3..1]".to_string()).unwrap()],
         );
 
-        assert!(transform.transform_one(event).is_none());
+        assert!(transform_one(&mut transform, event).is_none());
     }
 
     #[test]
@@ -308,7 +309,7 @@ mod tests {
             vec![Substring::new("only[10..11]".to_string()).unwrap()],
         );
 
-        assert!(transform.transform_one(event).is_none());
+        assert!(transform_one(&mut transform, event).is_none());
     }
 
     #[test]
@@ -322,6 +323,6 @@ mod tests {
             vec![Substring::new("only[..11]".to_string()).unwrap()],
         );
 
-        assert!(transform.transform_one(event).is_none());
+        assert!(transform_one(&mut transform, event).is_none());
     }
 }
