+++ src/transforms/lua/v2/mod.rs
@@ -1,17 +1,15 @@
-pub use vector_core::event::lua;
-
 use crate::{
     config::{DataType, CONFIG_PATHS},
     event::Event,
     internal_events::{LuaBuildError, LuaGcTriggered},
-    transforms::{
-        util::runtime_transform::{RuntimeTransform, Timer},
-        Transform,
-    },
+    transforms::Transform,
 };
 use serde::{Deserialize, Serialize};
 use snafu::{ResultExt, Snafu};
 use std::path::PathBuf;
+pub use vector_core::event::lua;
+use vector_core::transforms::runtime_transform::RuntimeTransform;
+use vector_core::transforms::runtime_transform::Timer;
 
 #[derive(Debug, Snafu)]
 pub enum BuildError {
