+++ src/transforms/split.rs
@@ -195,7 +195,9 @@ mod tests {
         let parser = parser.as_function();
 
         let metadata = event.metadata().clone();
-        let result = parser.transform_one(event).unwrap().into_log();
+        let mut buf = Vec::with_capacity(1);
+        parser.transform(&mut buf, event);
+        let result = buf.pop().unwrap().into_log();
         assert_eq!(result.metadata(), &metadata);
         result
     }
