--- src/transforms/util/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(any(feature = "transforms-lua"))]
-pub mod runtime_transform;
